requirements.txt
anyio==4.12.1
APScheduler==3.10.4
certifi==2026.1.4
charset-normalizer==3.4.4
elastic-transport==8.17.1
elasticsearch==8.11.0
exceptiongroup==1.3.1
greenlet==3.3.1
h11==0.16.0
httpcore==1.0.9
httpx==0.25.2
idna==3.11
numpy==1.26.4
packaging==26.0
pandas==2.1.3
psycopg2-binary==2.9.9
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-telegram-bot==20.7
pytz==2023.3
requests==2.31.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.23
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
urllib3==2.6.3
pymysql==1.1.0
cryptography==41.0.7
